merged chicken_breast and chicken_legs together
Query: in the Poultry-derived Food class, should the instances 'Chicken Breast' and 'Chicken legs' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Chicken breast and chicken legs are semantic and ontologic duplicates as they belong to the same meat cut category, which is chicken.
meaning that instances chicken_breast and chicken_legs are semantically identical.
Query cache in files:
- /kgfiller/data/cache-c5eb93fbf4f4150d3cd48f806d9aeb800c3a3da932c3cfa90d041d7d4b1ee2db.yml