instances chicken and chicken_stock were NOT merged together
Query: in the Poultry-derived Food class, should the instances 'Chicken stock' and 'Chicken' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Chicken stock and chicken are semantic and ontologic duplicates. They are both referring to the same thing, which is chicken.
meaning that instances chicken and chicken_stock are semantically identical.
instances not merged cause at least one was already merged previously.
Query cache in files:
- /kgfiller/data/cache-55c0310393f65ac48d3676c5c570b7aeed31c811cb4bea8dfd9ceef929ab8b4f.yml