instances chicken and chicken_thighs were NOT merged together
Query: in the Poultry-derived Food class, should the instances 'Chicken thighs' and 'Chicken' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No.

Chicken thighs and 'Chicken' are not semantic and ontologic duplicates. Although they are related concepts, they are distinct entities. Chicken thighs are a specific part of the chicken, while "Chicken" is a broader category of poultry.
meaning that instances chicken and chicken_thighs are different.
Query cache in files:
- /kgfiller/data/cache-aadbe87f38da3fce5bb83466dcc9b19aeddbcb3a9ebe02854748046b6cd56779.yml