instances chicken and chicken_legs were NOT merged together
Query: in the Poultry-derived Food class, should the instances 'Chicken legs' and 'Chicken' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes

Chicken legs are a part of the chicken carcass, so they are ontological duplicates of the chicken class.
meaning that instances chicken and chicken_legs are semantically identical.
instances not merged cause at least one was already merged previously.
Query cache in files:
- /kgfiller/data/cache-f0f1e6c3876225bf8cd96b9d860747768e443093d75f44b0fd1aea767d857733.yml