instances chicken and chicken_drumsticks were NOT merged together
Query: in the Poultry-derived Food class, should the instances 'Chicken drumsticks' and 'Chicken' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

The instances 'Chicken drumsticks' and 'Chicken' are semantic and ontologic duplicates. They refer to the same concept, which is chicken.
meaning that instances chicken and chicken_drumsticks are semantically identical.
instances not merged cause at least one was already merged previously.
Query cache in files:
- /kgfiller/data/cache-0c64e866cf96152a60eaec18919afedfbf23ee463f0b3ecacffe6a09d71e400a.yml