instances chicken_legs and chicken_drumsticks were NOT merged together
Query: in the Poultry-derived Food class, should the instances 'Chicken drumsticks' and 'Chicken legs' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Chicken drumsticks and chicken legs are semantic and ontologic duplicates, as they are both part of the same anatomical structure, namely the leg of a chicken.
meaning that instances chicken_legs and chicken_drumsticks are semantically identical.
instances not merged cause at least one was already merged previously.
Query cache in files:
- /kgfiller/data/cache-0884e6e001a94e1e2e69240ada50e6b0df5f9777e79f1e58a38e16d56f940ad9.yml