instances skim_milk and coconut_milk were NOT merged together
Query: in the Milk class, should the instances 'Coconut milk' and 'Skim Milk' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No.

Coconut milk and skim milk are not semantic and ontologic duplicates. They are two different types of milk with different nutritional profiles.
meaning that instances skim_milk and coconut_milk are different.
Query cache in files:
- /kgfiller/data/cache-ee228bd4433ff7069cd9ddfe0b9c7b5a82839b579ec80e7e96de71b984b82bde.yml