merged swordfish and fish_broth together
Query: in the Edible Fish class, should the instances 'Fish broth' and 'Swordfish' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Fish broth and Swordfish are semantic and ontologic duplicates. Fish broth is made of Swordfish.
meaning that instances swordfish and fish_broth are semantically identical.
Query cache in files:
- /kgfiller/data/cache-98d252f312d393083e9b04890136f42265584d788a6ac01d3ed50bc2c034d7e1.yml