merged swordfish and flying_fish_fillets together
Query: in the Edible Fish class, should the instances 'Flying Fish Fillets' and 'Swordfish' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	yes

The instances 'Flying Fish Fillets' and 'Swordfish' are semantic and ontologic duplicates as they both belong to the same taxonomic category, which is the species category of swordfish.
meaning that instances swordfish and flying_fish_fillets are semantically identical.
Query cache in files:
- /kgfiller/data/cache-31f6b42fd9133ef667627820956cf569f21b20d3a3d26b76d9e7faa9d449dbc3.yml