instances __crickets____crick and __grasshoppers____grass were NOT merged together
Query: in the Honey class, should the instances '**Crickets:** Crick' and '**Grasshoppers:** Grass' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No.

Crickets and Grasshoppers are two different species of insects, therefore they are not semantic and ontologic duplicates.
meaning that instances __crickets____crick and __grasshoppers____grass are different.
Query cache in files:
- /kgfiller/data/cache-6cd56099bc50063ffdbd2160b8e71b025600ac6fce2e9ca865fc2ed4fc6600ec.yml