merged instances_of_the_bread_class and bread together
Query: in the Bread class, should the instances 'Bread' and 'Instances of the Bread class' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	yes

The instances 'Bread' and 'Instances of the Bread class' are semantic and ontologic duplicates. They refer to the same concept.
meaning that instances instances_of_the_bread_class and bread are semantically identical.
Query cache in files:
- /kgfiller/data/cache-43b196992c0595f90d48360ad24720e9b09a3880effeaa55ad9e30490c3e7dcb.yml