merged instances_of_the_bread_class and breadcrumbs together
Query: in the Bread class, should the instances 'Breadcrumbs' and 'Instances of the Bread class' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

The instances 'Breadcrumbs' and 'Instances of the Bread class' are semantic and ontologic duplicates. They both refer to the same concept, which is the instances of the Bread class. Therefore, they should be merged together.
meaning that instances instances_of_the_bread_class and breadcrumbs are semantically identical.
Query cache in files:
- /kgfiller/data/cache-288280767cbeb19edc2b8b242ba27b40d9a29a300ed3c2c6cfd4802310a29c7e.yml