instances instances_of_the_bread_class and whole_wheat_bread were NOT merged together
Query: in the Bread class, should the instances 'Instances of the Bread class' and 'Whole Wheat Bread' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No

The instances 'Instances of the Bread class' and 'Whole Wheat Bread' are not semantic and ontologic duplicates, as they represent different concepts:

* **Instances of the Bread class**: This is a group of instances that belong to the Bread class. It includes all instances of bread, regardless of their specific type or ingredients.
* **Whole Wheat Bread**: This is a specific type of bread made from whole wheat flour. It is an instance of the Bread class and a subclass of the Whole Wheat Bread class.

Therefore, the two instances are not semantic or ontologic duplicates, as they represent different concepts.
meaning that instances instances_of_the_bread_class and whole_wheat_bread are different.
Query cache in files:
- /kgfiller/data/cache-fc16707b79af7acce5f1dbe780b1064c30184561ff02d24c99479881167cded3.yml