instances bread and bread_crumbs were NOT merged together
Query: in the Bread class, should the instances 'Bread crumbs' and 'Bread' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

The instances 'Bread crumbs' and 'Bread' are semantic and ontologic duplicates. They are referring to the same thing, which is bread.
meaning that instances bread and bread_crumbs are semantically identical.
instances not merged cause at least one was already merged previously.
Query cache in files:
- /kgfiller/data/cache-1a756634baf6ae37a49077d2cddb82f3896d6af77bd7e412b78f39808ada5528.yml