instances breadcrumbs and whole_wheat_bread were NOT merged together
Query: in the Bread class, should the instances 'Breadcrumbs' and 'Whole Wheat Bread' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes

Breadcrumbs and Whole Wheat Bread are semantic and ontologic duplicates as they both refer to the same type of food, which is bread.
meaning that instances breadcrumbs and whole_wheat_bread are semantically identical.
instances not merged cause at least one was already merged previously.
Query cache in files:
- /kgfiller/data/cache-fdee20387e4fdaea5e5dd6162e28fdff0e492aaffc8257d6938599fc5c603397.yml