merged flour and flour_tortilla together
Query: in the Bread class, should the instances 'Flour tortilla' and 'Flour' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Flour tortilla and Flour are semantic and ontologic duplicates, as they refer to the same entity: flour.
meaning that instances flour and flour_tortilla are semantically identical.
Query cache in files:
- /kgfiller/data/cache-7f7216c8ede5c2afae227a28ab418de00fbcd2392a6cd121e98bb85afc5189a1.yml