instances tortillas and vanilla_extract were NOT merged together
Query: in the Bread class, should the instances 'Tortillas' and 'Vanilla extract' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No.

Tortillas and Vanilla extract are not semantic or ontologic duplicates. Tortillas are a type of bread, while Vanilla extract is a flavoring.
meaning that instances tortillas and vanilla_extract are different.
Query cache in files:
- /kgfiller/data/cache-f7c73548350d16a668dfeee061af12c08279e1c8d622f6aa9b1bec621bc25300.yml