instances jasmine_rice and rice were NOT merged together
Query: in the Rice class, should the instances 'Jasmine Rice' and 'Rice' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Jasmine Rice and Rice are semantic and ontologic duplicates. They refer to the same food item, which is rice.
meaning that instances jasmine_rice and rice are semantically identical.
instances not merged cause at least one was already merged previously.
Query cache in files:
- /kgfiller/data/cache-5ddfef074bdf821085916f1cee32b58bfa3947dacefba5fb8f043af86fa4c82f.yml