instances long_grain_rice and whole_grains were NOT merged together
Query: in the Rice class, should the instances 'Long-Grain Rice' and 'Whole grains' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No.

Long-Grain Rice and Whole grains are not semantic and ontologic duplicates. They are different types of grains.
meaning that instances long_grain_rice and whole_grains are different.
Query cache in files:
- /kgfiller/data/cache-66191d18c9b7ef2f1cbd3b2ec1730043e4ddc89d309d20ed5c6c45bb097729ff.yml