instances wild_rice and sushi_rice were NOT merged together
Query: in the Rice class, should the instances 'Sushi Rice' and 'Wild Rice' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

The instances 'Sushi Rice' and 'Wild Rice' are semantic and ontologic duplicates, as they refer to the same type of rice, which is Wild Rice.
meaning that instances wild_rice and sushi_rice are semantically identical.
instances not merged cause at least one was already merged previously.
Query cache in files:
- /kgfiller/data/cache-313a0b2be3f5caf563d0248b87d46d7c19429f893c77e95acb1974cceeac336c.yml