merged lemon and lemon_zest together
Query: in the Edible Fruit class, should the instances 'Lemon zest' and 'Lemon' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

Lemon zest and lemon are semantic and ontologic duplicates. Lemon zest is a part of the lemon fruit.
meaning that instances lemon and lemon_zest are semantically identical.
Query cache in files:
- /kgfiller/data/cache-2e69ff438a85242e9066b8f28d8dda7ca6cdbfb434d07bf77e39b5bb206b8fee.yml