instances apple_filling and apple_crust were NOT merged together
Query: in the Edible Fruit class, should the instances 'Apple crust' and 'Apple filling' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	No.

The instances 'Apple crust' and 'Apple filling' are not semantic and ontologic duplicates, as they represent different parts of the apple.
meaning that instances apple_filling and apple_crust are different.
Query cache in files:
- /kgfiller/data/cache-9439a4cb974d2583c2470b13f8f26a6ceac0096f33c0bfedac4c3f4877179385.yml