instances apple_filling and fillings_of_choice were NOT merged together
Query: in the Edible Fruit class, should the instances 'Apple filling' and 'Fillings of choice' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	Yes.

The instances 'Apple filling' and 'Fillings of choice' are semantic and ontologic duplicates. They both refer to the same concept, which is the fillings of fruits, specifically the filling of an apple.
meaning that instances apple_filling and fillings_of_choice are semantically identical.
instances not merged cause at least one was already merged previously.
Query cache in files:
- /kgfiller/data/cache-cf82a90e378ec2386f5443dec43ad5dc19e168c1e58b670084277c183a3830f2.yml